+++ src/script/bindings/mod.rs
@@ -5,5 +5,6 @@ pub mod helpers;
 pub mod http;
 pub mod ntp;
 pub mod ping;
+pub mod redis;
 pub mod tcp;
 pub mod tls;